pyproject.toml
[tool.poetry]
name = "site.mahsummary"
version = "0.1.0"
description = "Website for the orgpedia/mahsummary dataset"
authors = ["mukundesh <mukundesh@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.8.1 <4.0"
pyright = "^1.1.355"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
jinja2 = "^3.1.3"
more-itertools = "^10.2.0"
ruff = "^0.3.3"
lunr = "^0.7.0.post1"
python-dateutil = "^2.9.0.post0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
lint.select = ["E", "F", "W", "I001"]
lint.ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["I001"] # fix only isort issues
lint.unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".nox",
    ".ruff_cache",
    ".tox",
    "__pypackages__",
    "dist",
    "node_modules",
    "venv",
    ".bak",
    "input",
    "output",
    "logs",
    "conf",
    "export",
    "models",
    "documents",
    "data_packages",
    "websites",
    "page_images",
    "html",
]
lint.per-file-ignores = {}

# Same as Black.
line-length = 100

# Allow unused variables when underscore-prefixed.
#dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.8
target-version = "py38"

[tool.pyright]
reportAttributeAccessIssue = false